chore: Adjust Skia for SystemThemeHelper support
