+++ src/Uno.UI.Runtime.Skia.Wpf/WPF/WpfHost.cs
@@ -6,6 +6,8 @@ using System.Windows;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using Uno.Foundation.Extensibility;
+using Uno.Helpers.Theming;
+using Uno.UI.Runtime.Skia.Wpf.WPF.Extensions.Helper.Theming;
 using Windows.Graphics.Display;
 using WinUI = Windows.UI.Xaml;
 using WpfApplication = System.Windows.Application;
@@ -22,7 +24,7 @@ namespace Uno.UI.Skia.Platform
 		{
 			ApiExtensibility.Register(typeof(Windows.UI.Core.ICoreWindowExtension), o => new WpfUIElementPointersSupport(o));
 			ApiExtensibility.Register(typeof(Windows.UI.ViewManagement.IApplicationViewExtension), o => new WpfApplicationViewExtension(o));
-			ApiExtensibility.Register(typeof(WinUI.IApplicationExtension), o => new WpfApplicationExtension(o));
+			ApiExtensibility.Register(typeof(ISystemThemeHelperExtension), o => new WpfSystemThemeHelperExtension(o));
 			ApiExtensibility.Register(typeof(IDisplayInformationExtension), o => new WpfDisplayInformationExtension(o));
 			ApiExtensibility.Register(typeof(Windows.ApplicationModel.DataTransfer.DragDrop.Core.IDragDropExtension), o => new WpfDragDropExtension(o));
 		}
