+++ src/Uno.UWP/Helpers/Theming/ISystemThemeHelperExtension.cs
@@ -1,0 +1,7 @@
+﻿namespace Uno.Helpers.Theming
+{
+	internal interface ISystemThemeHelperExtension
+	{
+		SystemTheme GetSystemTheme();
+	}
+}
