+++ src/Uno.UWP/Helpers/Theming/SystemThemeHelper.Android.cs
@@ -1,4 +1,6 @@
-﻿using Android.Content.Res;
+﻿#nullable enable
+
+using Android.Content.Res;
 using Android.OS;
 
 namespace Uno.Helpers.Theming
