+++ src/Uno.UWP/Helpers/Theming/SystemThemeHelper.macOS.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿#nullable enable
+
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
