+++ src/Uno.UWP/Helpers/Theming/SystemThemeHelper.wasm.cs
@@ -1,4 +1,6 @@
-﻿using System;
+﻿#nullable enable
+
+using System;
 using Uno.Extensions;
 using Uno.Foundation;
 using Uno.Logging;
